ATRT support for Windows (with Cygwin)

+++ storage/ndb/test/run-test/atrt-gather-result.sh
@@ -6,33 +6,37 @@ mkdir -p result
 cd result
 rm -rf *
 
-while [ $# -gt 0 ]
-do
-  rsync -a --exclude='BACKUP' --exclude='ndb_*_fs' "$1" .
-  shift
-done
+if uname | grep -iq cygwin; then
+  while [ $# -gt 0 ]
+  do
+    SAVE_IFS=$IFS
+    IFS=":"
+    declare -a ARR=($1)
+    IFS=$SAVE_IFS
+    DIR=`dirname "${ARR[1]}"`
+    REMOTE_DIR=`cygpath -u $DIR`
+    HOST="${ARR[0]}"
+    rsync -a --exclude='BACKUP' --exclude='ndb_*_fs' "$HOST:$REMOTE_DIR" .
+    shift
+  done
+else
+  while [ $# -gt 0 ]
+  do
+    rsync -a --exclude='BACKUP' --exclude='ndb_*_fs' "$1" .
+    shift
+  done
 
-# Cygwin doesn't like the rest of this program
-if uname -o > /dev/null 2>&1
-then
-  if [ `uname -o` = "Cygwin" ]
+  #
+  # clean tables...not to make results too large
+  #
+  lst=`find . -name '*.frm'`
+  if [ "$lst" ]
   then
-    exit 0
-  fi
-fi
-
-#
-# clean tables...not to make results too large
-#
-lst=`find . -name '*.frm'`
-if [ "$lst" ]
-then
-    for i in $lst
-    do
-	basename=`echo $i | sed 's!\.frm!!'`
-	if [ "$basename" ]
-	then
+ 	  basename=`echo $i | sed 's!\.frm!!'`
+	  if [ "$basename" ]
+	  then
 	    rm -f $basename.*
-	fi
-    done
+	  fi
+  fi
+
 fi
